update cosmovisor install script

+++ utils/cosmovisor_install.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
 
 # read arguments
-while getopts n:v:s:h flag; do
+while getopts b:h:s:v flag; do
   case "${flag}" in
-  n) BINARY_NAME=$OPTARG ;;
+  b) BINARY_NAME=$OPTARG ;;
   h) BINARY_HOME=$OPTARG ;;
   v) BINARY_VERSION=$OPTARG ;;
   s) SERVICE_NAME=$OPTARG ;;
@@ -11,8 +11,19 @@ while getopts n:v:s:h flag; do
   esac
 done
 
-# install cosmovisor binary
+if [ -z "$SERVICE_NAME" ]; then
+  SERVICE_NAME=BINARY_NAME
+fi
+
+if [ -z "$BINARY_VERSION" ]; then
+  echo "ERROR: binary version is undefined"
+  exit
+fi
+
 COSMOVISOR_VERSION="v1.3.0"
+COSMOVISOR_DIR="$HOME/$BINARY_HOME/cosmovisor"
+
+# install cosmovisor binary
 if [ -z "$(which cosmovisor)" ]; then
   cd || return
   rm -rf cosmos-sdk
@@ -20,20 +31,16 @@ if [ -z "$(which cosmovisor)" ]; then
   cd cosmos-sdk || return
   git checkout "cosmovisor/$COSMOVISOR_VERSION"
   make cosmovisor
-  mv cosmovisor/cosmovisor "$HOME/go/bin/cosmovisor"
+  mkdir -p "$COSMOVISOR_DIR"
+  mv cosmovisor/cosmovisor "$COSMOVISOR_DIR"
 fi
 
 # setup directories
-COSMOVISOR_DIR="$HOME/$BINARY_HOME/cosmovisor"
-if [ -z "$BINARY_VERSION" ]; then
-  mkdir -p "$COSMOVISOR_DIR/genesis/bin"
-  cp "$(which $BINARY_NAME)" "$HOME/$BINARY_HOME/cosmovisor/genesis/bin"
-  ln -s "$COSMOVISOR_DIR/genesis" "$COSMOVISOR_DIR/current"
-else
-  mkdir -p "$COSMOVISOR_DIR/upgrades/$BINARY_VERSION/bin"
-  cp "$(which $BINARY_NAME)" "$HOME/$BINARY_HOME/cosmovisor/$BINARY_VERSION/bin"
-  ln -s "$COSMOVISOR_DIR/upgrades/$BINARY_VERSION" "$COSMOVISOR_DIR/current"
-fi
+mkdir -p "$COSMOVISOR_DIR/genesis/bin"
+mkdir -p "$COSMOVISOR_DIR/upgrades/$BINARY_VERSION/bin"
+cp "$(which $BINARY_NAME)" "$COSMOVISOR_DIR/genesis/bin"
+cp "$(which $BINARY_NAME)" "$COSMOVISOR_DIR/upgrades/$BINARY_VERSION/bin"
+ln -s "$COSMOVISOR_DIR/upgrades/$BINARY_VERSION" "$COSMOVISOR_DIR/current"
 
 # update service file
 sudo tee "/etc/systemd/system/$SERVICE_NAME.service" > /dev/null << EOF
@@ -48,7 +55,7 @@ Restart=always
 RestartSec=3
 LimitNOFILE=10000
 Environment="DAEMON_NAME=$BINARY_NAME"
-Environment="DAEMON_HOME=$BINARY_HOME"
+Environment="DAEMON_HOME=$HOME/$BINARY_HOME"
 Environment="DAEMON_ALLOW_DOWNLOAD_BINARIES=true"
 Environment="DAEMON_RESTART_AFTER_UPGRADE=true"
 Environment="DAEMON_LOG_BUFFER_SIZE=512"
@@ -58,7 +65,6 @@ Environment="UNSAFE_SKIP_BACKUP=true"
 WantedBy=multi-user.target
 EOF
 
-# reload service files
 sudo systemctl daemon-reload
 sudo systemctl restart $SERVICE_NAME
 sudo journalctl -u $SERVICE_NAME -f -o cat --no-hostname
